implemented dinamic root scene switch  with buttons
