+++ client/src/main/scala/cml/ClientMain.scala
@@ -7,19 +7,15 @@ import javafx.fxml.FXMLLoader
 import javafx.scene.{Parent, Scene}
 import javafx.stage.Stage
 
-/**
-  * @author Monica Gondolini,Filippo Portolani
- */
-
 class ClientMain extends Application{
 
   override def start(primaryStage: Stage): Unit = {
     val rootParent: Parent = FXMLLoader.load(getClass.getClassLoader.getResource(AuthenticationWindow.path))
 
-    val scene : Scene = new Scene(rootParent, AuthenticationWindow.width, AuthenticationWindow.height)
+    val scene : Scene = new Scene(rootParent)
 
     ViewSwitchConfig.scenes = Map(
-      AuthenticationWindow.path -> FXMLLoader.load(getClass.getClassLoader.getResource(AuthenticationWindow.path)),
+      AuthenticationWindow.path -> rootParent,
       VillageWindow.path -> FXMLLoader.load(getClass.getClassLoader.getResource(VillageWindow.path)),
       BattleWindow.path -> FXMLLoader.load(getClass.getClassLoader.getResource(BattleWindow.path)),
       ArenaWindow.path -> FXMLLoader.load(getClass.getClassLoader.getResource(ArenaWindow.path))
